support opening and closing delimiters
For optimal Transcribe compatibility, conflicts with Markdown syntax
must be avoided. Markdown uses GREATER-THAN SIGN (>) at the beginning
of each line of a <blockquote>. This conflicts with that character's
use to signify doctest input, as in the following example:

    //# K :: a -> b -> a
    //.
    //. This is a paragraph.
    //.
    //. > This is a block quotation.
    //. > It should not be evaluated!
    //.
    //. ```javascript
    //. > K (true) (42)
    //. true
    //. ```

This commit adds --opening-delimiter and --closing-delimiter options to
the command-line interface, enabling the use of Markdown's <blockquote>
syntax in source files with delimited doctest blocks.

The following options would resolve the conflict in the example above:

    $ doctest --prefix . \
    .         --opening-delimiter '```javascript' \
    .         --closing-delimiter '```' \
    .         -- path/to/file.js