Implement support for doctesting ESM files
This commit adds a new choice to the 'module' option: 'esm'. When
running doctest with Node version 9 or up available, setting 'module'
to 'esm' allows for the documentation comments to be embedded in
ECMAScript modules, and use 'import' to load dependencies.

The approach included in this commit has the following consequences:

1. The CLI transparently switches between ESM and non-ESM enabled
   based on the Node version running it. This means the CLI is
   fully backwards compatible.
2. The programmatic version also transparently switches between
   ESM and non-ESM depending on whether it's loaded via import
   or via require.
3. The programmatic version has a breaking change, in that its
   primary function returns a Promise now.

Co-Authored-By: David Chambers <dc@davidchambers.me>